Conference request over HTTP (#2188)
* ref: Inline onMucMemberLeft (has nothing to do with "moderator" and is broken).

* fix: Use isFocus instead of just reading the resource.

* ref: Inline onParticipantLeft.

* ref: Add _ to private function names.

* ref: Remove redundant function.

* ref: Initialize focusComponent once.

* ref: Remove two copies of "options".

* ref: Simplify the focusJid flow.

* ref: Rename focusCompoonent to targetJid.

* fix: Only retry on "invalid session" errors

We used to fire FOCUS_DISCONNECTED (which doesn't have anything to do
with anything disconnecting) and then retrying. This is not necessary
because FOCUS_DISCONNECTED is a fatal error and the client schedules
a reload of it's own (and we want the client reload as opposed to the
conference-request retry in case e.g. the shard has changed).

* ref: Remove use of visitorFocus.

* ref: Remove graceful shutdown handling (never used in backend).

* feat: Support conference request over HTTP.

* ref: eslint cleanup.

* fix: Add max timeout 2 minutes.