feat(GlobalOnErrorHandler) completely drop it
All modern browsers implement window.addEventListener for the 'error'
and 'unhandledrejection' listeners. Thus there is zero benefit from
using the legacy `onerror` handler here and rolling multiple handlers
ourselves.

There is an old config option `enableWindowOnErrorHandler` which was
never whitelisted in Jitsi Meet, so I'm 100% sure nobody is using it.

Even if someone was, migrarting to the above is simply better in every
possible way.