fix the deadlock
JingleSessionPC queue is paused until the first offer/answer.

If JitsiConference has a track operation on it's queue,
before session offer/answer, they will never execute,
because the track operation blocks starting the JingleSession's
queue.

It's okay to skip track operations on JingleSessionPC, because
it will pick up the local track on the offer/answer.