fix(statistics) initialize LocalStatsCollector on Statistics.init
Creating the AudioContext at import time means a warning line is printed
in the JS console even before the user has a chance of initializing the
library, which is weird.

There is no harm in initializing it in init since it's the first thing
the user needs to do in order to use the library.