Expose `track.conference` field in types
Local and remote Jitsi tracks expose a reference to the
`JitsiConference` through `track.conference`. This was not available in
the TypeScript definitions.

Local tracks are initialized with `conference = null` but will be set
later, assuming the caller adds the track to a conference. Remote tracks
are always associated with a conference.