fix(JitsiConference) drop confusing log line
The reason is already logged when the function is called, and we really
don't care about the result either way.

In addition, there are legit reasons to reject the incoming session, for
example because the user has disabled P2P, so the "invalid" message in
the logs was already confusing.