fix(SDP) Do not filter out SSRC 'cname' attributes. (#2695)
* fix(SDP) Do not filter out SSRC 'cname' attributes.
Chrome doesn't generate the a=ssrc line with 'msid' attribute when the track is removed from pc and a renegotiation is triggered. As a result a source-remove is sent to the p2p peer when it shouldn't be causing issues on the receiver when the same source is added back to the conference. Fixes toggleCamera on mobile browser not working as expected.