fix(VP9): Disable for FF.
Disables VP9 for FF since the simulcast  is disabled with FF flag by default. Introduces a few fixes so that if the FF flag is enabled simulcast works. This should also help us to be ready to enable VP9 once the  FF config flag for simulcast for VP9 is enabled by default.