fix(remoteControl): Failure to start
It appears that when the gDM flow is used we do not set properly the sourceId of the shared stream.  When the user has multiple screens we use the sourceId to identify the display that is currently shared. Therefore in electron we can't identify the display that is used and we fail.